requirements.txt
Flask==2.3.3
Pillow==9.5.0
FontTools==4.42.1
waitress==2.1.2
